endpoint/requirements.txt
flask
pyjwt
gunicorn
# redis
